requirements.txt
pandas
matplotlib.pyplot
yfinance
streamlit
datetime
